# tara-system/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TARA System API
  description: |
    汽车智能威胁分析与风险评估系统 API 文档
    
    ## 概述
    TARA System 提供完整的威胁分析和风险评估功能，符合 ISO/SAE 21434 标准。
    
    ## 认证
    所有 API 请求需要在 Header 中携带 JWT Token:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: TARA System Team
  license:
    name: MIT
servers:
  - url: http://localhost:8001/api/v1
    description: Development server
  - url: https://tara.example.com/api/v1
    description: Production server
...
paths:
...
  /threats:
    get:
      tags: [Threats]
      summary: 获取威胁列表
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
  /threats/analyze:
    post:
      tags: [Threats]
      summary: STRIDE威胁分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: integer
                asset_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '202':
          description: 分析任务已启动
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []